🤖 test: create user mock and use in tests
