+++ src/app/user/useCases/GetUserById/__tests__/GetUserByIdController.spec.ts
@@ -1,13 +1,6 @@
 import { app } from '../../../../main';
 import request from 'supertest';
-import { ICreateUserDTO } from '../../CreateUser/CreateUserDTO';
-
-const mockUser: ICreateUserDTO = {
-  name: 'Lucas Torres',
-  email: 'lucastorres@make2u.com.br',
-  password: 'liven123',
-  birthday: new Date(),
-};
+import { mockUser } from '../../../../shared/TestMock';
 
 describe('Unit test: Get User [Controller]', () => {
   let id: string;
@@ -20,7 +13,9 @@ describe('Unit test: Get User [Controller]', () => {
 
   it('should be able to get an existent user by id', async () => {
     const response = await request(app).get(`/users/${id}`);
+    const user = JSON.parse(response.text);
 
     expect(response.status).toBe(200);
+    expect(user).toHaveProperty('id');
   });
 });
